Fantasy Basketball Automation
Based on the Vegas line for team wins at the beginning of the season, some friends and I drafted which teams we though would win more or less than the predicted win totals. Whoever has the highest score above the vegas line at the end of the season wins the league. This code scraps the web to update win and loss totals, while the excel file is used to constantly update the score. Feel free to use it with you and your friends!

Code created using brute force, not the prettiest but gets the job done.